bank-ai-assistant/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
matplotlib
xlsxwriter
requests
faker
asyncpg
httpx
